SSE-3294 Connect sheets functionality to frontend
